requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.1
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.5
matplotlib==3.9.1
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==2.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paho-mqtt==2.1.0
pillow==10.4.0
protobuf==4.25.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.14.0
six==1.16.0
sounddevice==0.4.7
